+++ integrations/azure-devops/main.py
@@ -63,7 +63,7 @@ from integration import (
     AzureDevopsTeamResourceConfig,
     AzureDevopsWorkItemResourceConfig,
     AzureDevopsTestRunResourceConfig,
-    AzureDevopsCodeCoverageResourceConfig,
+    AzureDevopsBuildCodeCoverageResourceConfig,
     AzureDevopsPullRequestResourceConfig,
     AzureDevopsAdvancedSecurityResourceConfig,
     AzureDevopsRepositoryResourceConfig,
@@ -430,18 +430,18 @@ async def resync_test_runs(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
         yield test_runs
 
 
-@ocean.on_resync(Kind.CODE_COVERAGE)
-async def resync_code_coverages(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
+@ocean.on_resync(Kind.BUILD_CODE_COVERAGE)
+async def resync_build_code_coverages(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     azure_devops_client = AzureDevopsClient.create_from_ocean_config()
     selector = cast(
-        AzureDevopsCodeCoverageResourceConfig, event.resource_config
+        AzureDevopsBuildCodeCoverageResourceConfig, event.resource_config
     ).selector
     coverage_config = selector.code_coverage
 
-    async for coverages in azure_devops_client.generate_code_coverages(
+    async for coverages in azure_devops_client.generate_build_code_coverages(
         coverage_config,
     ):
-        logger.info(f"Fetched {len(coverages)} code coverage entries")
+        logger.info(f"Fetched {len(coverages)} build code coverage entries")
         yield coverages